removed historical dependency on junk in base.h fixed a bunch of  signedness and unused variable warnings fixed some data types that really shouldn't have been signed

+++ include/core/recommendation_list.h
@@ -85,7 +85,7 @@ public:
     int sigma2Distance = -1;
     int listSizeMultiplier = 200;
     int ficFavouritesCutoff = 0;
-    int resultLimit = 0;
+    size_t resultLimit = 0;
     uint16_t ratioCutoff = std::numeric_limits<uint16_t>::max();
 
     double quadraticDeviation = -1;
